fyiona/requirements.txt
asyncpg
Django
django-filter
djangorestframework
firebase_admin
Markdown
mixer
Pillow
pyjwt
psycopg2-binary
pytest-cov 
pytest-django
python-dotenv
pytz
sqlparse
django-phonenumber-field[phonenumberslite]
django-rest-passwordreset
django-cors-headers
requests